fix:
30 minutes

+++ modules/networker/src/main/java/ru/vachok/networker/sysinfo/AppConfigurationLocal.java
@@ -73,7 +73,7 @@ public interface AppConfigurationLocal extends Runnable {
     }
 
     default void schedule(Runnable runnable, int timeInMinPerion) {
-        schedule(runnable, 0, (int) TimeUnit.MINUTES.toMillis(timeInMinPerion));
+        schedule(runnable, 0, TimeUnit.MINUTES.toMillis(timeInMinPerion));
     }
 
     default void schedule(Runnable runnable, long timeFirstRun, long period) {
